Retry jobs that failed because they were not defined (yet)
The worker may fail to load some jobs because the job class does not
exist at the time of execution. This can happen on upgrades when the
worker is still running the previous version while a migration is
enqueuing jobs defined in the new version. The worker fails
deserializing them because the job class can not be loaded.

This commit adds an initializer to retry all jobs failing due to
NameError on restart so that they can be picked up once the migration is
over and the worker gets restarted.