[49409] Handle excluded statuses from totals in update job
The job has been updated to excluded work packages having a status with
`excluded_from_totals` set to true. The journal entry has valid values
so that the cause renders correctly.

The job is not yet triggered properly when the "excluded from totals"
attribute of a status is changed in the administration.