Compute totals even if all children are excluded
There was a bug in the way the excluded work packages were affecting the
query which updates the totals, especially regarding looking at the
`MAX(generations)` value. When considering work packages after filtering
out excluded work packages, the parent is alone, and its
`MAX(generations)` is 0, which is interpreted like it is a leaf work
package and leading to totals being unset, which is not correct.

Now the check for `MAX(generations) > 0` is done outside of the
filtering to prevent this bug from happening.