fix signaling on the mail_alert_sent
It is now possible to distinguish by reason whether the mail_alert_sent field is set to true, nil or false.

This didn`t suffice though to have the immediate_reminders user preference handled for mentioning mails. Ideally, the preference would be handled
in the CreateFromModelService to have the field set correctly right away. This does not work with the current structure though so the alternative
of bailing on actually sending the mail is used. The property is then set to nil by that service upon identifying the user as not wanting
to receive mails