[#55961] Define the project lists query for activated storages
https://community.openproject.org/work_packages/55961

This also defines a basic table in the view, so it can be tested.
I've choosen not to go with a new query filter for projects belonging
to a storage because it seemed to be more work than necessary and not
provide any further benefits, since for the API there already is
such a filter (which also seems to be tailored to the needs of the
API. It would not have worked for the table because the table should
include archived projects as well).