[55997] Add duration_format setting with hours_only by default
https://community.openproject.org/wp/55997

The `duration_format` setting can take two values: "hours_only" or
"days_and_hours". The default value is "hours_only".

The `duration_format` setting is used to determine how durations are
displayed in the frontend and in some exports.

It can be changed in Administration > Calendar and Dates > Working days
and hours.