Remember back_url/origin as RelayState in saml
When SAML uses the HTTP-POST binding, we lose the saved back_url as the
session cookie is no longer sent due to SameSite=Lax.

We also don't want to widen SameSite to None, so we need another way to
keep this information. SAML has RelayState that we can use for storing this state.

https://community.openproject.org/work_packages/55188