Fix flickering spec spec/features/projects/navigation_spec.rb:45
Can be seen failing on https://github.com/opf/openproject/actions/runs/9775000755/job/26984486804?pr=16029

It seems to happen because the component is clicked before it's able to properly handle it: the subscription to this.projects$ is not complete yet.

Adding a ReplaySubject for synchronization seems to be enough to fix it.