Fix flaky feature tests involving timeline
Failing spec is ./spec/features/work_packages/table/context_menu/context_menu_spec.rb:68
It was failing in https://github.com/opf/openproject/actions/runs/9853177518/job/27203169350

The test is opening the context menu of a work package by right clicking
in the work package list. But the `refreshView()` of the timeline is
called and this closes the context menu before the test assertion run,
making the test fail. Waiting for network idle when checking if the
timeline is open fixes it.