refactor: move logic around to make methods less complex
The check for invalid values is done once instead of being done in each
update method, and the value is updated only if the attribute can be
derived. An attribute can only be derived if the user did not set it and
some value from the 2 other attributes have changed.