Fix writable_attributes when redefining attributes in child classes
There were cases were the writable parameter of a parent contract would
take precedence of the child contract writable parameter when the parent
condition is a lambda: a child could say `writable: true` and it would
still be the parent `writable: -> { false }` that would "win" because
when reducing and evaluating `writable_conditions`, only lambda are
considered.

This commit fixes the issue by adding extensive tests and putting both
lambdas AND true/false values in the `writable_conditions` hash.