[#55968] Improve spec
Instead of only checking if the page contains the project folder types
anywhere, it should also be ensured that they are in the expected table row.

An alternative to this would have been

```
expect(page).to have_text("#{project.name} Automatically managed", normalize_ws: true)
expect(page).to have_text("#{archived_project.name} No specific folder", normalize_ws: true)
```

But to make it explicit that we want to check that the contents of a
table row match our expectations, I added a new "within" method for
selecting table rows.