[#56780] Add copy to clipboard button for newly created API token
https://community.openproject.org/work_packages/56780/activity

Actually, the whole layout of the dialog was changed, because of UX
implications of having a copy to clipboard button and to make the
user more aware that he has something to do.

From a technical side, I switched from using a blankslate to using
a flex layout, because the blankslate would enforce the text to be
centered and the input label and warning are left aligned and do not
look good when being centered. I also tried having them below the
blankslate (so the blankslate would only have an icon and an heading)
but then there's a large margin between the heading and the rest of
the content.