[#57028] Reinstate dynamic project storage deletion via turbo stream
This reverts commit f3a0db2f149dfef639e88aa4500421f49e2549b2.

Update the controller to respond with partial turbo streams instead of a full page redirect with turbo enabled. This caused DOM load issues with angular components, as Turbo would not break out of it's navigation frame- to remount the angular web components. The solution here is also used in other similar pages such as project attributes.