Fix flaky move_work_package_spec.rb spec
The spec was triggering the move work package job, then was immediately
making an assertion on the page current path. This assertion waits 4
seconds max, and it can be too few for the things being done:
- The job status dialog polls for status every 2 seconds.
- Then once it returns, it displays a successful message for another 2
  seconds before redirecting automatically
- The redirection goes to /work_package/:id, which redirects to
  /projects/:project_identifier/work_packages/:id/activity

Only then will the assertion be true. On CI, it often takes more than 4
seconds to get to this state which leads to false negatives.

To prevent that:
- Wait for the job status dialog to display its success message.
- Then wait for 2 seconds for the auto-redirect to trigger.

Clicking the link would be better but there is a bug about it. Once the
bug is fixed, the link can be clicked and that will reduce the test
execution time.