Fix date picker being blank and jumpy when appearing
It was in order to fix the flaky spec
`spec/features/admin/working_days_spec.rb:189`. The flakiness can be
forced by adding a `sleep 2` at the beginning of the test, just before
the `datepicker.open_modal!` call.

Not absolutely sure why so many forced change detection calls are
needed, but this way it seems to work. It seems to be related to having
a lot of nested templates, outlets and `if`s in templates which each
time need a new change detection to be reevaluated. But I'm not an
expert in this.

As a bonus there is no delay anymore in the drop modal initialization,
which should lead to more stable tests. For instance execution time of
`spec/features/admin/working_days_spec.rb:189` dropped from 17 secs to 7
secs.