Fix tests broken in previous commit
It was stupid to make an assertion on the selected day for the
datepicker in `#select_day` as the date picker is dismissed half of the
time, so the selected element is not there anymore to make assertions on
it.

Replaced with a direct assertion where it matters and reverted the
"safeguard".