[#57704] Resolve pending spec in `direct_ifc_upload_spec.rb` (#16664)
https://community.openproject.org/work_packages/57704

When the file input gets a file, it makes a request to the backend to
check for the file size. This request takes some time. The test was not
waiting for the request to complete before checking the form validity.

By adding a check on the validation message of the file input, we make
sure the test waits for the request to complete and the that form
validity check is correct.

---------

Co-authored-by: Christophe Bliard <c.bliard@openproject.com>