Remove `hostDirective` that already exist in parent class
It prevents the component from being displayed with the error "Uncaught
RuntimeError: NG0309: Directive WorkPackageIsolatedQuerySpaceDirective
matches multiple times on the same element. Directives can only match an
element once.".

Indeed, the same `hostDirective` was already defined in the parent
class. It prevented the query configuration modal to be displayed when
adding a "table of related work packages" in a type form configuration,
which was detected with test
`spec/features/types/form_configuration_query_spec.rb` not passing
locally. Indeed, on a local development environment, the modal could not
be displayed.

What is strange is that it seemed to be working properly in CI and in
production: the modal could be displayed despite the hostDirective being
probably duplicated there too.