Set % Complete to 100% when status is closed depending on settings
If setting `percent_complete_on_status_closed` is set to `set_100p`,
then in work-based mode the % complete is set to 100% when the status is
closed. It can still be modified afterwards.

In status-based mode, this setting has no effect. The % complete is
still set to the status's default % complete value.