Fix flakiness in spec work_package_activity_spec.rb
Two different changes were made: one changing work, one changing
remaining work. In activities, the entry was created on the first
change, and then updated with new content on the second change.

When `within("activity-entry", text: admin.name)` was used, the first
entry could be returned depending on how fast the server responds. That
entry would then become stale when the entry is updated with the new
content, leading to the test failing with the following error: expected
to find visible list item with text "% Complete set to 95%" within
"<<ERROR>>", "<<ERROR>>", "<<ERROR>>", "<<ERROR>>", "<<ERROR>>",
"<<ERROR>>", which matched the selector but not all filters.

"<<ERROR>>" is what is returned when the element cannot return its text
content.

This happened in https://github.com/opf/openproject/actions/runs/10926176246/job/30329450428

This flakiness error was fixed by doing only one modification to the
work package, so that the entry is created once and never updated.