[56769] Migrate queries with timeline to Gantt module
https://community.openproject.org/wp/56769

The original migration `ChangeViewOfQueriesWithTimelineToGantt` was
fine, but it was applied too early: in OpenProject 13.2.0 the migration
would already have been run and it was still possible to create Gantt
queries inside the work packages module. Such queries were not migrated.

This commit runs the original migration again to ensure all queries
displayed as Gantt charts are displayed in the right module.