Display status previous name in edit form
If the status gets edited, and the name is changed, and the form has
errors, the edit form will be shown again. It needs to show the previous
name as this is the one being edited. It's most obvious when the status
name is emptied and submitted and the "Name can't be blank" error is
displayed.