[57251] Fix mail_alert_sent not being set correctly on update
https://community.openproject.org/wp/57251

When a notification exists for a work package journal for a reason other
than :mentioned, then its `mail_alert_sent` value is `nil`.

On a subsequent update of the work package which mentions the user, the
existing notification is updated. In this update, the `mail_alert_sent`
value must be changed from `nil` to `false` if the reason is :mentioned.
Without it, the immediate notification email for the mention would not
be sent.